\name{fibPkgWithRcpp-package}
\alias{fibPkgWithRcpp-package}
\alias{fibPkgWithRcpp}
\docType{package}
\title{\packageTitle{fibPkgWithRcpp}}
\description{\packageDescription{fibPkgWithRcpp}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{fibPkgWithRcpp}
  \packageIndices{fibPkgWithRcpp}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{fibPkgWithRcpp}
  
  Maintainer: \packageMaintainer{fibPkgWithRcpp}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
